Refactor `schemamd` and `functionmd` packages to `internal/`
