+++ README.md
@@ -158,8 +158,8 @@ The `validate` subcommand can be used to validate the provider website documenta
 |---------------------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | `InvalidDirectoriesCheck` | Checks for valid subdirectory structure and throws an error if an invalid Terraform Provider documentation subdirectory is found.                                                   |
 | `MixedDirectoriesCheck`   | Throws an error if both legacy documentation (`/website/docs`) and registry documentation (`/docs`) are found.                                                                      |
-| `NumberOfFilesCheck`      | Throws an error if the number of files in a directory is larger than the registry limit.                                                                                            |
-| `FileSizeCheck`           | Throws an error if the documentation file is above the registry storage limit.                                                                                                      |
+| `NumberOfFilesCheck`      | Throws an error if the number of files in a directory is larger than the public registry limit.                                                                                     |
+| `FileSizeCheck`           | Throws an error if the documentation file is above the public registry storage limit.                                                                                               |
 | `FileExtensionCheck`      | Throws an error if the extension of the given file is not a valid registry documentation extension.                                                                                 |
 | `FrontMatterCheck`        | Checks the YAML frontmatter of documentation for missing required fields or invalid fields.                                                                                         |
 | `FileMismatchCheck`       | Throws an error if the names/number of resources/datasources/functions in the provider schema does not match the names/number of files in the corresponding documentation directory |
@@ -326,12 +326,6 @@ or, if you can, submit a [pull-request](https://github.com/hashicorp/terraform-p
 Your help and patience is truly appreciated.
 
 ## Contributing
-
-## Go Compatibility
-
-This project follows the [support policy](https://golang.org/doc/devel/release.html#policy) of Go as its support policy. The two latest major releases of Go are supported by the project.
-
-Currently, that means Go **1.21** or later must be used when including this project as a dependency.
 
 ### License Headers
 All source code files in this repository (excluding autogenerated files like `go.mod`, prose, and files excluded in [.copywrite.hcl](.copywrite.hcl)) must have a license header at the top.
