+++ internal/functionmd/render_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	tfjson "github.com/hashicorp/terraform-json"
 
-	"github.com/hashicorp/terraform-plugin-docs/functionmd"
+	"github.com/hashicorp/terraform-plugin-docs/internal/functionmd"
 )
 
 func TestRenderArguments(t *testing.T) {
