+++ internal/functionmd/testdata/example_vararg.md
