+++ internal/provider/template.go
@@ -16,10 +16,10 @@ import (
 
 	tfjson "github.com/hashicorp/terraform-json"
 
-	"github.com/hashicorp/terraform-plugin-docs/functionmd"
+	"github.com/hashicorp/terraform-plugin-docs/internal/functionmd"
 	"github.com/hashicorp/terraform-plugin-docs/internal/mdplain"
+	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
 	"github.com/hashicorp/terraform-plugin-docs/internal/tmplfuncs"
-	"github.com/hashicorp/terraform-plugin-docs/schemamd"
 )
 
 const (
