+++ internal/schemamd/render_test.go
@@ -11,7 +11,8 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	tfjson "github.com/hashicorp/terraform-json"
-	"github.com/hashicorp/terraform-plugin-docs/schemamd"
+
+	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
 )
 
 func TestRender(t *testing.T) {
