+++ internal/schemamd/testdata/aws_acm_certificate.schema.json
