+++ internal/schemamd/testdata/aws_route_table_association.md
