+++ internal/schemamd/testdata/awscc_acmpca_certificate.md
