+++ internal/schemamd/testdata/awscc_logs_log_group.md
