+++ internal/schemamd/write_attribute_description.go
