+++ internal/schemamd/write_attribute_description_test.go
@@ -11,7 +11,7 @@ import (
 	tfjson "github.com/hashicorp/terraform-json"
 	"github.com/zclconf/go-cty/cty"
 
-	"github.com/hashicorp/terraform-plugin-docs/schemamd"
+	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
 )
 
 func TestWriteAttributeDescription(t *testing.T) {
