+++ internal/schemamd/write_block_type_description_test.go
@@ -9,7 +9,8 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	tfjson "github.com/hashicorp/terraform-json"
-	"github.com/hashicorp/terraform-plugin-docs/schemamd"
+
+	"github.com/hashicorp/terraform-plugin-docs/internal/schemamd"
 )
 
 func TestWriteBlockTypeDescription(t *testing.T) {
