+++ internal/schemamd/write_nested_attribute_type_description.go
